+++ engine/runtime/src/main/java/org/enso/interpreter/node/callable/resolver/HostMethodCallNode.java
@@ -59,6 +59,26 @@ public abstract class HostMethodCallNode extends Node {
      * Standard.Base.Data.Time.Date} and dispatching natively.
      */
     CONVERT_TO_DATE,
+    /**
+     * The method call should be handled by converting {@code self} to a {@code
+     * Standard.Base.Data.Time.Date_Time} and dispatching natively.
+     */
+    CONVERT_TO_ZONED_DATE_TIME,
+    /**
+     * The method call should be handled by converting {@code self} to a {@code
+     * Standard.Base.Data.Time.Date_Time} with a system Time_Zone and dispatching natively.
+     */
+    CONVERT_TO_DATE_TIME,
+    /**
+     * The method call should be handled by converting {@code self} to a {@code
+     * Standard.Base.Data.Time.Time_Of_Day} and dispatching natively.
+     */
+    CONVERT_TO_TIME_OF_DAY,
+    /**
+     * The method call should be handled by converting {@code self} to a {@code
+     * Standard.Base.Data.Time.Time_Zone} and dispatching natively.
+     */
+    CONVERT_TO_TIME_ZONE,
     /** The method call should be handled by dispatching through the {@code Any} type. */
     NOT_SUPPORTED;
 
@@ -70,7 +90,13 @@ public abstract class HostMethodCallNode extends Node {
      * @return true if one can directly pass this object to {@link InteropLibrary}
      */
     public boolean isInteropLibrary() {
-      return this != NOT_SUPPORTED && this != CONVERT_TO_TEXT && this != CONVERT_TO_DATE;
+      return this != NOT_SUPPORTED
+          && this != CONVERT_TO_TEXT
+          && this != CONVERT_TO_DATE
+          && this != CONVERT_TO_DATE_TIME
+          && this != CONVERT_TO_ZONED_DATE_TIME
+          && this != CONVERT_TO_TIME_OF_DAY
+          && this != CONVERT_TO_TIME_ZONE;
     }
   }
 
@@ -93,8 +119,20 @@ public abstract class HostMethodCallNode extends Node {
    */
   public static PolyglotCallType getPolyglotCallType(
       Object self, String methodName, InteropLibrary library) {
-    if (library.isDate(self) && !library.isTime(self)) {
-      return PolyglotCallType.CONVERT_TO_DATE;
+    if (library.isDate(self)) {
+      if (library.isTime(self)) {
+        if (library.isTimeZone(self)) {
+          return PolyglotCallType.CONVERT_TO_ZONED_DATE_TIME;
+        } else {
+          return PolyglotCallType.CONVERT_TO_DATE_TIME;
+        }
+      } else {
+        return PolyglotCallType.CONVERT_TO_DATE;
+      }
+    } else if (library.isTime(self)) {
+      return PolyglotCallType.CONVERT_TO_TIME_OF_DAY;
+    } else if (library.isTimeZone(self)) {
+      return PolyglotCallType.CONVERT_TO_TIME_ZONE;
     } else if (library.isString(self)) {
       return PolyglotCallType.CONVERT_TO_TEXT;
     } else if (library.isMemberInvocable(self, methodName)) {